<2020.12.31 HeeSang-Yun> 1. 보상 지급 부분에서 값이 정상적으로 맞지 않는 버그 Fix

+++ Common.h
@@ -58,9 +58,9 @@ typedef struct StReward_Result
     std::string ErrorCode;
     std::string ErrorMsg;
 
-    float		Total_Reward_Coin;
+    long double	Total_Reward_Coin;
 
-    float		Price;
+    long double	Price;
     __int64		Work_Time;
 
     bool		bState;
@@ -80,7 +80,7 @@ typedef struct StReward_Result
 
 typedef struct StReward_Info
 {
-    float Price;
+    long double Price;
     __int64 Work_Time;
 
     std::string	 Wallet_addr;
@@ -116,12 +116,12 @@ typedef struct StRewardList
 
 typedef struct StSumReward_Info
 {
-    __int64 Price;
+    long double Price;
     std::string	 Wallet_addr;
 
     StSumReward_Info()
     {
-        Price = 0;
+        Price = 0.0f;
         Wallet_addr.clear();
     }
 }SumReward_Info;
@@ -160,10 +160,10 @@ typedef struct StReReward
 
 typedef struct stBalanceInfo
 {
-    __int64 Balance;
+    long double Balance;
 
     stBalanceInfo()
     {
-        Balance = 0;
+        Balance = 0.0f;
     }
 }BalanceInfo;
